+++ src/eca/features/tools.clj
@@ -7,7 +7,8 @@
    [eca.features.tools.mcp :as f.mcp]
    [eca.features.tools.shell :as f.tools.shell]
    [eca.features.tools.util :as tools.util]
-   [eca.logger :as logger])
+   [eca.logger :as logger]
+   [eca.messenger :as messenger])
   (:import
    [java.util Map]))
 
@@ -29,21 +30,31 @@
           (when (get-in config [:nativeTools :shell :enabled])
             f.tools.shell/definitions))))
 
+(defn ^:private native-tools [db config]
+  (mapv #(select-keys % [:name :description :parameters])
+        (vals (native-definitions db config))))
+
 (defn all-tools
   "Returns all available tools, including both native ECA tools
    (like filesystem and shell tools) and tools provided by MCP servers."
   [db config]
-  (let [native-tools (concat
-                      []
-                      (mapv #(select-keys % [:name :description :parameters])
-                            (vals (native-definitions db config))))
-        mcp-tools (f.mcp/all-tools db)]
-    (concat
-     (mapv #(assoc % :origin :native) native-tools)
-     (mapv #(assoc % :origin :mcp) mcp-tools))))
+  (concat
+   (mapv #(assoc % :origin :native) (native-tools db config))
+   (mapv #(assoc % :origin :mcp) (f.mcp/all-tools db))))
 
 (defn call-tool! [^String name ^Map arguments db config]
   (logger/info logger-tag (format "Calling tool '%s' with args '%s'" name arguments))
   (if-let [native-tool-handler (get-in (native-definitions db config) [name :handler])]
     (native-tool-handler arguments {:db db :config config})
     (f.mcp/call-tool! name arguments db)))
+
+(defn init-servers! [db* messenger config]
+  (messenger/tool-server-updated messenger {:type :native
+                                            :name "ECA"
+                                            :status "running"
+                                            :tools (native-tools @db* config)})
+  (f.mcp/initialize-servers-async!
+   {:on-server-updated (fn [server]
+                         (messenger/tool-server-updated messenger (assoc server :type :mcp)))}
+   db*
+   config))