+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Include eca as a  server with tools.
+
 ## 0.8.1
 
 - Make generic tool server updates for eca native tools.